feat: support types only

+++ src/Generator.ts
@@ -4,7 +4,21 @@ import JSON5 from 'json5'
 import path from 'path'
 import request from 'request-promise-native'
 import { castArray, isEmpty, isFunction } from 'vtils'
-import { CategoryList, Config, ExtendedInterface, Interface, InterfaceList, Method, PropDefinition, RequestBodyType, RequestFormItemType, Required, ResponseBodyType, ServerConfig, SyntheticalConfig } from './types'
+import {
+  CategoryList,
+  CliConfig,
+  ExtendedInterface,
+  Interface,
+  InterfaceList,
+  Method,
+  PropDefinition,
+  RequestBodyType,
+  RequestFormItemType,
+  Required,
+  ResponseBodyType,
+  ServerConfig,
+  SyntheticalConfig,
+} from './types'
 import { getNormalizedRelativePath, jsonSchemaStringToJsonSchema, jsonSchemaToType, jsonToJsonSchema, mockjsTemplateToJsonSchema, propDefinitionsToJsonSchema, throwError } from './utils'
 import { JSONSchema4 } from 'json-schema'
 
@@ -19,12 +33,16 @@ export class Generator {
   /** 配置 */
   private config: ServerConfig[] = []
 
+  /** 是否只提取types */
+  private typesOnly: boolean = false
+
   /** { 项目标识: 分类列表 } */
   private projectIdToCategoryList: Record<string, CategoryList | undefined> = Object.create(null)
 
-  constructor(config: Config) {
+  constructor(cliConfig: CliConfig) {
+    this.typesOnly = cliConfig.typesOnly
     // config 可能是对象或数组，统一为数组
-    this.config = castArray(config)
+    this.config = castArray(cliConfig.serverConfigs)
   }
 
   async generate(): Promise<OutputFileList> {
@@ -52,6 +70,7 @@ export class Generator {
                           ...serverConfig,
                           ...projectConfig,
                           ...categoryConfig,
+                          typesOnly: this.typesOnly,
                           mockUrl: projectInfo.getMockUrl(),
                         }
                         syntheticalConfig.devUrl = projectInfo.getDevUrl(syntheticalConfig.devEnvName!)
@@ -63,7 +82,7 @@ export class Generator {
                         )
                         const categoryUID = `_${serverIndex}_${projectIndex}_${categoryIndex}`
                         const categoryCode = [
-                          [
+                          syntheticalConfig.typesOnly ? '' : [
                             `const mockUrl${categoryUID} = ${JSON.stringify(syntheticalConfig.mockUrl)} as any`,
                             `const devUrl${categoryUID} = ${JSON.stringify(syntheticalConfig.devUrl)} as any`,
                             `const prodUrl${categoryUID} = ${JSON.stringify(syntheticalConfig.prodUrl)} as any`,
@@ -119,7 +138,7 @@ export class Generator {
         const { content, requestFilePath } = outputFileList[outputFilePath]
 
         // 若 request 文件不存在，则写入 request 文件
-        if (!(await fs.pathExists(requestFilePath))) {
+        if (!(await fs.pathExists(requestFilePath)) && !this.typesOnly) {
           await fs.outputFile(
             requestFilePath,
             `${`
@@ -187,7 +206,7 @@ export class Generator {
         await fs.outputFile(
           outputFilePath,
           `${
-            [
+            this.typesOnly ? content.join('\n\n').trim() : [
               `/* tslint:disable */\n/* eslint-disable */`,
               `/**\n * **该文件由 yapi-to-typescript 自动生成，请勿直接修改！！！** \n */`,
               `import request from ${JSON.stringify(requestFileRelativePathWithoutExt)}`,
@@ -384,34 +403,36 @@ export class Generator {
       requestDataType.trim(),
       `\n/**\n * 接口 **${interfaceInfo.title}** 的 **返回类型**\n */`,
       responseDataType.trim(),
-      `\n/**\n * 接口 **${interfaceInfo.title}** 的 **请求函数**\n */`,
-      `export function ${requestFunctionName}(requestData${isRequestDataRequired ? '?' : ''}: ${requestDataTypeName}): Promise<${responseDataTypeName}> {`,
-      [
-        `  return request({`,
-        `    ...${requestFunctionName}.requestConfig,`,
-        `    ...parseRequestData(requestData)`,
-        `  } as any)`,
-      ].join('\n'),
-      `}`,
-      [
-        `\n/**\n * 接口 **${interfaceInfo.title}** 的 **请求配置**\n */`,
-        `${requestFunctionName}.requestConfig = Object.freeze({`,
-        `  mockUrl: mockUrl${categoryUID},`,
-        `  devUrl: devUrl${categoryUID},`,
-        `  prodUrl: prodUrl${categoryUID},`,
-        `  path: ${JSON.stringify(interfaceInfo.path)},`,
-        `  method: Method.${interfaceInfo.method},`,
-        `  requestBodyType: RequestBodyType.${interfaceInfo.method === Method.GET ? RequestBodyType.query : interfaceInfo.req_body_type || RequestBodyType.none},`,
-        `  responseBodyType: ResponseBodyType.${interfaceInfo.res_body_type},`,
-        `  dataKey: dataKey${categoryUID}`,
-        `} as RequestConfig<`,
-        `  ${JSON.stringify(syntheticalConfig.mockUrl)},`,
-        `  ${JSON.stringify(syntheticalConfig.devUrl)},`,
-        `  ${JSON.stringify(syntheticalConfig.prodUrl)},`,
-        `  ${JSON.stringify(interfaceInfo.path)},`,
-        `  ${JSON.stringify(syntheticalConfig.dataKey)}`,
-        `>)`,
-      ].join('\n'),
+      ...(syntheticalConfig.typesOnly ? [] : [
+        `\n/**\n * 接口 **${interfaceInfo.title}** 的 **请求函数**\n */`,
+        `export function ${requestFunctionName}(requestData${isRequestDataRequired ? '?' : ''}: ${requestDataTypeName}): Promise<${responseDataTypeName}> {`,
+        [
+          `  return request({`,
+          `    ...${requestFunctionName}.requestConfig,`,
+          `    ...parseRequestData(requestData)`,
+          `  } as any)`,
+        ].join('\n'),
+        `}`,
+        [
+          `\n/**\n * 接口 **${interfaceInfo.title}** 的 **请求配置**\n */`,
+          `${requestFunctionName}.requestConfig = Object.freeze({`,
+          `  mockUrl: mockUrl${categoryUID},`,
+          `  devUrl: devUrl${categoryUID},`,
+          `  prodUrl: prodUrl${categoryUID},`,
+          `  path: ${JSON.stringify(interfaceInfo.path)},`,
+          `  method: Method.${interfaceInfo.method},`,
+          `  requestBodyType: RequestBodyType.${interfaceInfo.method === Method.GET ? RequestBodyType.query : interfaceInfo.req_body_type || RequestBodyType.none},`,
+          `  responseBodyType: ResponseBodyType.${interfaceInfo.res_body_type},`,
+          `  dataKey: dataKey${categoryUID}`,
+          `} as RequestConfig<`,
+          `  ${JSON.stringify(syntheticalConfig.mockUrl)},`,
+          `  ${JSON.stringify(syntheticalConfig.devUrl)},`,
+          `  ${JSON.stringify(syntheticalConfig.prodUrl)},`,
+          `  ${JSON.stringify(interfaceInfo.path)},`,
+          `  ${JSON.stringify(syntheticalConfig.dataKey)}`,
+          `>)`,
+        ].join('\n'),
+      ]),
     ].join('\n')
   }
 }
